pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "growgrow"
version = "0.1.0"
description = "Portfolio tracker and analyzer for Tradernet"
requires-python = ">=3.11"
dependencies = [
    "tradernet-sdk>=2.0.0",
    "pandas>=2.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
growgrow = "growgrow.cli:main"

[tool.setuptools.packages.find]
include = ["growgrow*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
